requirements.txt
# MCP Framework
fastmcp>=0.4.0

# Document Parsing
pdfplumber>=0.10.0
pymupdf>=1.23.0
openpyxl>=3.1.0
pandas>=2.1.0

# Browser Automation
playwright>=1.40.0

# BigQuery / Data Lake
google-cloud-bigquery>=3.13.0
google-cloud-bigquery-storage>=2.24.0
db-dtypes>=1.2.0

# Anthropic (normalization LLM)
anthropic>=0.23.0

# Data Validation & Schemas
pydantic>=2.5.0
python-dateutil>=2.8.0

# Dashboard
streamlit>=1.29.0
plotly>=5.18.0
streamlit-aggrid>=0.3.4

# REST API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6   # file uploads
python-jose[cryptography]>=3.3.0  # JWT auth

# Multi-tenant Platform DB
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tenacity>=8.2.3        # retry logic for API calls
structlog>=23.2.0      # structured logging
httpx>=0.25.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
